Apply ES^ best practices to refactor the code
