+++ src/modules/obj-constructor.js
